requirements.txt
supabase>=2.5.0
requests>=2.32.0
python-dotenv>=1.0.1
pyserial>=3.5
numpy>=1.26.0
matplotlib>=3.8.0
